Closed #18.

+++ test/smart_ptr_test.cc
@@ -14,13 +14,13 @@ TEST(utils, smart_ptr_count) {
     auto ptr_2 = LogicXor::Init();
     auto ptr_3 = LogicXor::Init();
   }
-  EXPECT_EQ(size_t(3), SPIStaticDataMember::cache_container_.size());
+  EXPECT_EQ(3u, SPIStaticDataMember::cache_container_.size());
 }
 
 
 TEST(utils, smart_ptr_free) {
   auto ptr = LogicAnd::Init();
-  EXPECT_EQ(long(2), ptr.use_count());
+  EXPECT_EQ(2l, ptr.use_count());
   SPIStaticDataMember::FreeCache();
-  EXPECT_EQ(long(1), ptr.use_count());
+  EXPECT_EQ(1l, ptr.use_count());
 }
